requirements.txt
streamlit
google-generativeai
srt
google-genai
